Fold JobWatchService back into a node-scoped JobMonitoringService
With @ScopeOptional, one service expresses both halves: the ES-backed
finders opt in so any node answers them unscoped, while watch stays
instance-affine and is invoked with the scope recorded on
JobRun.nodeId. The separate JobWatchService, its TS mirror, and the
extracted JobRunAuthorizer fold away.
