+++ kinotic-js/workspace/packages/management-api/src/api/services/IJobMonitoringService.ts
@@ -1,14 +1,19 @@
 import { MANAGEMENT_API_ZONE } from '@/api/PlatformZones'
 import type { IKinotic, IServiceProxy, Page, Pageable } from '@kinotic-ai/core'
+import type { Observable } from 'rxjs'
 import type { JobRun } from '@/api/model/grind/JobRun'
 import type { TaskRecord } from '@/api/model/grind/TaskRecord'
+import type { JobRunEvent } from '@/api/model/grind/events/JobRunEvent'
 
 /**
- * Read access to grind job runs for the authenticated participant: an organization or
- * application participant sees the runs its organization owns, a system participant sees
- * every run. A run's TaskRecords are its task ledger - every discovered task has a record,
- * PENDING until it starts executing. The live event stream of a run is served by
- * IJobWatchService on the node named by JobRun.nodeId.
+ * Access to grind job runs for the authenticated participant: an organization or application
+ * participant sees the runs its organization owns, a system participant sees every run. A
+ * run's TaskRecords are its task ledger - every discovered task has a record, PENDING until
+ * it starts executing.
+ *
+ * The finders read shared state, so any node answers them; watch serves a run's live event
+ * stream, which exists only in the executing process, so it is invoked with the node id
+ * recorded on JobRun.nodeId as its scope.
  */
 export interface IJobMonitoringService {
 
@@ -30,6 +35,15 @@ export interface IJobMonitoringService {
      * @param pageable the page of records to return
      */
     findTasks(jobRunId: string, pageable: Pageable): Promise<Page<TaskRecord>>
+
+    /**
+     * Opens a live view of a job run the participant may view, replaying every JobRunEvent
+     * emitted since the run started and continuing until the run terminates. Completes
+     * without emissions when the run is not currently executing on the node.
+     * @param nodeId the id of the node executing the run, from JobRun.nodeId
+     * @param jobRunId the id of the run to watch
+     */
+    watch(nodeId: string, jobRunId: string): Observable<JobRunEvent>
 }
 
 export class JobMonitoringService implements IJobMonitoringService {
@@ -50,5 +64,9 @@ export class JobMonitoringService implements IJobMonitoringService {
 
     public findTasks(jobRunId: string, pageable: Pageable): Promise<Page<TaskRecord>> {
         return this.serviceProxy.invoke('findTasks', [jobRunId, pageable])
+    }
+
+    public watch(nodeId: string, jobRunId: string): Observable<JobRunEvent> {
+        return this.serviceProxy.invokeStream('watch', [jobRunId], nodeId)
     }
 }
