+++ kinotic-js/workspace/packages/management-api/src/index.ts
@@ -86,7 +86,6 @@ export * from '@/api/services/INamedQueriesDefinitionService'
 export * from '@/api/services/IMigrationService'
 export * from '@/api/services/IDataInsightsService'
 export * from '@/api/services/IJobMonitoringService'
-export * from '@/api/services/IJobWatchService'
 export * from '@/api/services/IMemberService'
 export * from '@/api/services/IInviteEmailTemplateService'
 export * from '@/api/services/IOAuthApprovalService'
