+++ website/content/02.platform/11.reference/02.grind-jobs.md
@@ -173,16 +173,22 @@ Every run persists a `JobRun` — name, version, owner, status, timestamps, erro
 
 ## Watching a run remotely
 
-`JobService.watchRun(jobRunId)` attaches to a run executing in this process, replaying it from the start; it returns an empty stream for a run that has finished or never started, and subscribing never starts one. `JobMonitoringService` publishes that stream, and the records behind it, to remote callers in the management-api zone:
+`JobService.watchRun(jobRunId)` attaches to a run executing in this process, replaying it from the start; it returns an empty stream for a run that has finished or never started, and subscribing never starts one. `JobMonitoringService` publishes that stream, and the records behind it, to remote callers in the management-api zone. Each node publishes its own instance, scoped by its node id: the finders are `@ScopeOptional`, so any instance answers an unscoped call, while `watch` must be invoked with the scope recorded on `JobRun.nodeId` — the node whose process holds the live stream:
 
 ```java
 @Publish
 public interface JobMonitoringService {
 
+    @Scope
+    String nodeId();
+
+    @ScopeOptional
     Future<Page<JobRun>> findJobRuns(Pageable pageable);
 
+    @ScopeOptional
     Future<JobRun> findJobRun(String jobRunId);
 
+    @ScopeOptional
     Future<Page<TaskRecord>> findTasks(String jobRunId, Pageable pageable);
 
     Flux<JobRunEvent> watch(String jobRunId);
